pyproject.toml
[project]
name = "typelefter"
version = "0.1.0"
description = "Typelefter is a free and open-source desktop app for prose and novel writing. Built with Python, PySide, and SQLite, it offers a clean, typewriter-inspired interface for structuring, drafting, and compiling manuscripts. Markdown-based, Git-integrated, and GPLv3-licensed to stay free for writers everywhere."
authors = [
    {name = "vijayv95",email = "vijayv95@gmail.com"}
]
license = {text = "GNU General Public License Version 3, 29 June 2007"}
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pyside6 (>=6.10.0,<7.0.0)",
    "gitpython (>=3.1.45,<4.0.0)",
    "markdown (>=3.9,<4.0)"
]

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
